Remove User Management GUI. Resolves T6881.
gnome-system-tools is being considered legacy software and is being gracefully and gradually deprecated. While I understand it is not optiminal to have non-functioning user management, we are actively working with the developer of an [alternative tool](https://github.com/zhuyaliang/user-admin) to provide a modern replacement and will ship it when it is considered production ready.